chore: Vendor fractional-indexing as a package
